[modify]move snabbdom example.
