+++ example/components/snabbdom/Routes/Bar.js
