+++ example/components/snabbdom/Routes/Foo.js
