+++ example/components/snabbdom/Routes/index.js
